requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.7.0
django-model-utils==5.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
